refactor: restructure views to use ModelViewSet + routers , restructure the views serializer tasks and urls in base app folder
