+++ core/settings.py
@@ -27,7 +27,7 @@ environ.Env.read_env(os.path.join(BASE_DIR, ".env"))
 # See https://docs.djangoproject.com/en/5.2/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = env("SECURITY_KEY")
+SECRET_KEY = env("SECRET_KEY")
 DEBUG = env.bool("DEBUG", default=True)
 
 # SECURITY WARNING: don't run with debug turned on in production!
@@ -212,16 +212,16 @@ from celery.schedules import crontab
 
 CELERY_BEAT_SCHEDULE = {
     "test-task-every-30-sec":{
-    "task": "accounts.api.tasks.test_beat_task",
+    "task": "accounts.tasks.test_beat_task",
     "schedule": 30.0,
     "args": (),
   },
   "update_stock_prices_every_minute":{
-    "task": "accounts.api.tasks.update_stock_prices",
+    "task": "accounts.tasks.update_stock_prices",
     "schedule": 60.0, 
   },
   "daily_trade_report":{
-      "task":"accounts.api.tasks.generate_daily_report",
+      "task":"accounts.tasks.generate_daily_report",
       "schedule": crontab(hour=0, minute=0),
   },
 }
